Add `BaseIntegrationEnv`.
