+++ Tests/ZkSync2Tests/Integration/Account/CustomAccountTests.swift
@@ -7,7 +7,7 @@
 
 import XCTest
 
-class CustomAccountTests: XCTestCase {
+class CustomAccountTests: BaseIntegrationEnv {
     
     override func setUpWithError() throws {
