+++ Tests/ZkSync2Tests/ZkSyncWalletIntegrationTests.swift
@@ -110,7 +110,7 @@ class ZkSyncWalletIntegrationTests: XCTestCase {
         DispatchQueue.global().async { [weak self] in
             guard let self = self else { return }
             
-            let web3 = try! Web3.new(ZKSyncWeb3RpcIntegrationTests.L1NodeUrl)
+            let web3 = try! Web3.new(BaseIntegrationEnv.L1NodeUrl)
             
             let amount = Web3.Utils.parseToBigUInt("9", units: .eth)!
